Add ToBeAdded section (awful name, but sticking  with Laser Maze conventions)
