+++ src/app/widgets.rs
@@ -1,3 +1,1 @@
-pub mod bank;
 pub mod cell;
-pub mod grid;
